[Concurrency] Add an environment variable to validate unchecked continuation usage.
When enabled, we track all active unchecked continuations in a global set, and fatal error if one is called twice.

This commit also removes `swift_Concurrency_fatalError` in favor of exporting `swift::fatalError` from libswiftCore as SPI so that libswift_Concurrency can use it. This gives us better error messages and crash reporting for fatal errors in libswift_Concurrency.